refactor: split out common/git
